fix: launch real Mac agent app in managed E2Es
The macOS managed-agent path should use the same app bundle that native
Mac users run, not a parallel headless executable. Reusing the app build
and quit plumbing keeps the E2E harness aligned with production while
still preserving the Go daemon path for non-macOS managed agents.
